Move choose, choose_mut and shuffle from Rng to SliceExt
